refactor(element3): move Message to src/components
