--- packages/element3/packages/message/index.js
@@ -1,1 +1,0 @@
-export { Message } from './Message.js'
