+++ packages/element3/packages/theme-chalk/src/index.scss
@@ -40,7 +40,6 @@
 @import "./menu.scss";
 @import "./menu-item.scss";
 @import "./menu-item-group.scss";
-@import "./message.scss";
 @import "./message-box.scss";
 @import "./notification.scss";
 @import "./page-header.scss";
@@ -77,5 +76,6 @@
 @import "./button.scss";
 @import "./container.scss";
 @import "./link.scss";
+@import "./message.scss";
 @import "./switch.scss";
 @import "./tag.scss";
