+++ packages/element3/src/components/Message/index.js
@@ -1,0 +1,1 @@
+export { Message } from './src/Message.js'
