+++ packages/element3/src/components/Message/tests/Message.spec.js
@@ -1,4 +1,4 @@
-import { Message } from '../Message.js'
+import { Message } from '../src/Message.js'
 import { h } from 'vue'
 describe('Message', () => {
   afterEach(async () => {
